+++ lib/awful/init.lua
@@ -8,8 +8,11 @@
 
 require("awful._compat")
 
-return
-{
+local deprecated = {
+    ewmh = true
+}
+
+local ret = {
     client = require("awful.client");
     completion = require("awful.completion");
     layout = require("awful.layout");
@@ -30,11 +33,21 @@ return
     wibox = require("awful.wibox");
     startup_notification = require("awful.startup_notification");
     tooltip = require("awful.tooltip");
-    ewmh = require("awful.ewmh");
+    permissions = require("awful.permissions");
     titlebar = require("awful.titlebar");
     rules = require("awful.rules");
     popup = require("awful.popup");
     spawn = require("awful.spawn");
 }
+
+-- Lazy load deprecated modules to reduce the numbers of loop dependencies.
+return setmetatable(ret,{
+    __index = function(_, key)
+        if deprecated[key] then
+            rawset(ret, key, require("awful."..key))
+        end
+        return rawget(ret, key)
+    end
+})
 
 -- vim: filetype=lua:expandtab:shiftwidth=4:tabstop=8:softtabstop=4:textwidth=80
