+++ tests/test-awful-client.lua
@@ -145,9 +145,9 @@ table.insert(steps, function()
     -- This should still be the case
     assert(c2.active)
 
-    original_count = #awful.ewmh.generic_activate_filters
+    original_count = #awful.permissions.generic_activate_filters
 
-    awful.ewmh.add_activate_filter(function(c)
+    awful.permissions.add_activate_filter(function(c)
         if c == c1 then return false end
     end)
 
@@ -161,13 +161,13 @@ table.insert(steps, function()
     assert(c2.active)
 
     -- Test the remove function
-    awful.ewmh.remove_activate_filter(function() end)
+    awful.permissions.remove_activate_filter(function() end)
 
-    awful.ewmh.add_activate_filter(awful.ewmh.generic_activate_filters[1])
+    awful.permissions.add_activate_filter(awful.permissions.generic_activate_filters[1])
 
-    awful.ewmh.remove_activate_filter(awful.ewmh.generic_activate_filters[1])
+    awful.permissions.remove_activate_filter(awful.permissions.generic_activate_filters[1])
 
-    assert(original_count == #awful.ewmh.generic_activate_filters)
+    assert(original_count == #awful.permissions.generic_activate_filters)
 
     c1:emit_signal("request::activate", "i_said_so")
